app/requirements.txt
flask
prometheus_client
dnspython